pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxtract"
version = "0.1.0"
description = "Audio to speaker-attributed transcript — extract who said what, when"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "voxtract contributors" }]
keywords = ["speech", "transcription", "diarization", "asr", "speaker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
stt = ["qwen-asr>=0.0.6", "torch>=2.0"]
speaker = ["pyannote.audio>=3.3"]
all = ["voxtract[stt,speaker]"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "numpy>=1.26",
]

[project.scripts]
voxtract = "voxtract.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/voxtract"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
